fix table header
