+++ app/client/src/widgets/wds/WDSTableWidget/component/StaticTable.tsx
@@ -8,8 +8,8 @@ import type { ReactElementType } from "react-window";
 import "simplebar-react/dist/simplebar.min.css";
 import type { ReactTableColumnProps, TableSizes } from "./Constants";
 import { MULTISELECT_CHECKBOX_WIDTH, TABLE_SCROLLBAR_WIDTH } from "./Constants";
-import type { TableColumnHeaderProps } from "./header/TableColumnHeader";
-import TableColumnHeader from "./header/TableColumnHeader";
+import type { TableColumnHeaderProps } from "./TableHeader/TableColumnHeader";
+import TableColumnHeader from "./TableHeader/TableColumnHeader";
 import { TableBody } from "./TableBody";
 
 type StaticTableProps = TableColumnHeaderProps & {
