+++ app/client/src/widgets/wds/WDSTableWidget/component/TableHeader/PageNumberInput.tsx
@@ -1,40 +1,7 @@
-import React, { useCallback, useEffect } from "react";
-import styled from "styled-components";
-import { NumericInput, Keys } from "@blueprintjs/core";
+import { Keys } from "@blueprintjs/core";
+import { TextInput } from "@design-system/widgets";
+import React, { useCallback, useEffect, useState } from "react";
 import { EventType } from "constants/AppsmithActionConstants/ActionConstants";
-
-const PageNumberInputWrapper = styled(NumericInput)`
-  &&& input {
-    box-shadow: none;
-    border: 1px solid var(--color-bd-neutral);
-    background: var(--color-bg);
-    box-sizing: border-box;
-    width: 24px;
-    height: 24px;
-    line-height: 24px;
-    padding: 0 !important;
-    text-align: center;
-    font-size: 12px;
-    border-radius: var(--border-radius-elevation-3);
-
-    &:disabled {
-      opacity: var(--opacity-disabled);
-    }
-  }
-
-  &&&.bp3-control-group > :only-child {
-    border-radius: 0;
-  }
-
-  & input:hover:not(:disabled) {
-    border-color: var(--color-bd-neutral-hover) !important;
-  }
-
-  & input:focus {
-    border-color: unset !important;
-    box-shadow: 0 0 0 1px var(--color-bd-focus) !important;
-  }
-`;
 
 const MIN_PAGE_COUNT = 1;
 
@@ -44,7 +11,7 @@ function PageNumberInputComponent(props: {
   updatePageNo: (pageNo: number, event?: EventType) => void;
   disabled: boolean;
 }) {
-  const [pageNumber, setPageNumber] = React.useState(props.pageNo || 0);
+  const [pageNumber, setPageNumber] = useState(props.pageNo || "0");
 
   useEffect(() => {
     setPageNumber(props.pageNo || 0);
@@ -78,23 +45,20 @@ function PageNumberInputComponent(props: {
   );
 
   return (
-    <PageNumberInputWrapper
-      buttonPosition="none"
-      clampValueOnBlur
+    <TextInput
       className="t--table-widget-page-input"
-      disabled={props.disabled}
-      max={props.pageCount || 1}
-      min={1}
+      isDisabled={props.disabled}
       onBlur={handleUpdatePageNo}
+      onChange={(value) => {
+        setPageNumber(value);
+      }}
       onKeyDown={(e: any) => {
         if (e.keyCode === Keys.ENTER) {
           handleUpdatePageNo(e);
         }
       }}
-      onValueChange={(value: number) => {
-        setPageNumber(value);
-      }}
-      value={pageNumber}
+      size="small"
+      value={pageNumber.toString()}
     />
   );
 }
