--- app/client/src/widgets/wds/WDSTableWidget/component/header/banner/index.tsx
@@ -1,12 +1,0 @@
-import React, { memo } from "react";
-import type { AddNewRowBannerType } from "./AddNewRowBanner";
-import { AddNewRowBanner } from "./AddNewRowBanner";
-
-export interface BannerPropType extends AddNewRowBannerType {}
-
-export const Banner = memo((props: BannerPropType) => (
-  <AddNewRowBanner
-    disabledAddNewRowSave={props.disabledAddNewRowSave}
-    onAddNewRowAction={props.onAddNewRowAction}
-  />
-));
